Refactor into pages/components
