+++ src/App.js
@@ -1,10 +1,9 @@
 import {BrowserRouter as Router, Switch, Route, Link} from 'react-router-dom'
 import {useState, useEffect} from 'react'
 import {supabase} from './supabaseClient'
-import Auth from './Auth'
-import Account from './Account'
-import Channel from './Channel'
-import NoMatch from './404'
+import Auth from './components/auth'
+import Account from './pages/account'
+import NoMatch from './pages/404'
 
 export default function Home() {
 	const [session, setSession] = useState(null)
@@ -24,13 +23,11 @@ export default function Home() {
 		<Router>
 			<nav>
 				<Link to="/">Home</Link>
+				<button onClick={() => supabase.auth.signOut()}>Sign Out</button>
 			</nav>
 			<Switch>
 				<Route exact path="/">
 					<Account key={session.user.id} session={session}></Account>
-				</Route>
-				<Route path="/channel">
-					<Channel key={session.user.id} session={session}></Channel>
 				</Route>
 				<Route path="*">
 					<NoMatch />
