+++ src/components/auth.js
@@ -1,5 +1,5 @@
 import {useState} from 'react'
-import {supabase} from './supabaseClient'
+import {supabase} from '../supabaseClient'
 
 export default function Auth() {
 	const [loading, setLoading] = useState(false)
